+++ lensmemoriae/models/ir_ui_view.py
@@ -12,7 +12,6 @@ class IrUiView(models.Model):
         if not user.has_group("lensmemoriae.group_moderator"):
             return arch
         prefs = {
-            "action_open_scrap_wizard": user.lensmemoriae_show_scrap,
             "action_clear_images": user.lensmemoriae_show_clear_all,
             "action_download_pending": user.lensmemoriae_show_download,
         }