+++ lensmemoriae/models/res_users.py
@@ -4,10 +4,6 @@ from odoo import fields, models
 class ResUsers(models.Model):
     _inherit = "res.users"
 
-    lensmemoriae_show_scrap = fields.Boolean(
-        string="Show Scrap button",
-        default=True,
-    )
     lensmemoriae_show_clear_all = fields.Boolean(
         string="Show Clear All button",
         default=False,