+++ lensmemoriae/wizards/__init__.py
@@ -1,5 +1,4 @@
 from . import reject_reason
 from . import location_picker
 from . import bulk_tag
-from . import scrap_wizard
 from . import arxiu_fons_wizard